Seems that SQL_MODE='' it's not enough for MariaDB v10
In some installations the error *"Incorrect date value: '0' for column 'the_date'"* occurs.

For example in MariaDB 10.3.12 the next statement fails with the mentioned error:
```SQL
SET SESSION SQL_MODE = ''; call get_products_detail('5','0','','0','1','0');
-- '5' is id of a provider
```

Therefore even if the SQL_MODE='' is forced in `php/inc/database.php` the execution fails.

It is proposed to use `varchar` instead of `date` in the procedures in which problems are detected.